# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BOC Study Notebook API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notebooks/{id}/study-guides:
    get:
      operationId: listStudyGuides
      tags: [studyguides]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StudyGuide" }
    post:
      operationId: generateStudyGuide
      tags: [studyguides]
      summary: AI-generate a study guide from notebook material
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudyGuideInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudyGuide" }
  /study-guides/{id}:
    get:
      operationId: getStudyGuide
      tags: [studyguides]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudyGuide" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deleteStudyGuide
      tags: [studyguides]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }

  # ---- Audio Overviews (TTS) ----
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: Curated study resources (filtered by topic/domain)
      parameters:
        - { name: topicId, in: query, required: false, schema: { type: integer } }
        - { name: domainId, in: query, required: false, schema: { type: integer } }
        - { name: kind, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Resource" }
    post:
      operationId: createResource
      tags: [resources]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResourceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resource" }
  /resources/{id}:
    delete:
      operationId: deleteResource
      tags: [resources]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    ApiError:
      type: object
      properties: { error: { type: string } }
      required: [error]
...
    StudyGuide:
      type: object
      required: [id, notebookId, title, format, content, createdAt]
      properties:
        id: { type: integer }
        notebookId: { type: integer }
        title: { type: string }
        format:
          type: string
          enum: [outline, summary, qa, mindmap]
        content: { type: string, description: "Markdown" }
        createdAt: { type: string, format: date-time }
    StudyGuideInput:
      type: object
      required: [format]
      properties:
        format:
          type: string
          enum: [outline, summary, qa, mindmap]
        focus: { type: string }
...
    Resource:
      type: object
      required: [id, title, url, kind, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        url: { type: string }
        kind: { type: string, enum: [article, video, podcast, book, paper, guideline, other] }
        provider: { type: ["string", "null"] }
        topicId: { type: ["integer", "null"] }
        domainId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        saved: { type: boolean }
        createdAt: { type: string, format: date-time }
    ResourceInput:
      type: object
      required: [title, url, kind]
      properties:
        title: { type: string, minLength: 1 }
        url: { type: string, format: uri }
        kind: { type: string, enum: [article, video, podcast, book, paper, guideline, other] }
        provider: { type: string }
        topicId: { type: integer }
        domainId: { type: integer }
        notes: { type: string }
...