# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BOC Study Notebook API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputKind } from "./resourceInputKind";

export interface ResourceInput {
  /** @minLength 1 */
  title: string;
  url: string;
  kind: ResourceInputKind;
  provider?: string;
  topicId?: number;
  domainId?: number;
  notes?: string;
}